feat: added kyber as provider to prevent multiple instances
