+++ package.json
@@ -33,6 +33,7 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.451.0",
     "@aws-sdk/s3-request-presigner": "^3.451.0",
+    "@dashlane/pqc-kem-kyber512-node": "^1.0.0",
     "@hapi/joi": "^17.1.1",
     "@internxt/inxt-js": "^2.0.8",
     "@internxt/lib": "^1.2.0",
@@ -63,6 +64,7 @@
     "crypto": "^1.0.1",
     "crypto-js": "^4.2.0",
     "geoip-lite": "^1.4.8",
+    "hash-wasm": "^4.12.0",
     "helmet": "^7.1.0",
     "mariadb": "^3.2.2",
     "multer-s3": "^3.0.1",
