+++ src/modules/user/user.module.ts
@@ -47,6 +47,7 @@ import { FeatureLimitModule } from '../feature-limit/feature-limit.module';
 import { WorkspacesModule } from '../workspaces/workspaces.module';
 import { SequelizeWorkspaceRepository } from '../workspaces/repositories/workspaces.repository';
 import { UserNotificationTokensModel } from './user-notification-tokens.model';
+import { AsymmetricEncryptionModule } from '../../externals/asymmetric-encryption/asymmetric-encryption.module';
 
 @Module({
   imports: [
@@ -74,6 +75,7 @@ import { UserNotificationTokensModel } from './user-notification-tokens.model';
     SecurityModule,
     forwardRef(() => FeatureLimitModule),
     forwardRef(() => WorkspacesModule),
+    AsymmetricEncryptionModule,
   ],
   controllers: [UserController],
   providers: [
