+++ src/modules/user/user.usecase.ts
@@ -50,11 +50,7 @@ import { ShareUseCases } from '../share/share.usecase';
 import { AvatarService } from '../../externals/avatar/avatar.service';
 import { SequelizePreCreatedUsersRepository } from './pre-created-users.repository';
 import { PreCreateUserDto } from './dto/pre-create-user.dto';
-import {
-  decryptMessageWithPrivateKey,
-  encryptMessageWithPublicKey,
-  generateNewKeys,
-} from '../../lib/assymetric-encryption/openpgp';
+
 import { aes } from '@internxt/lib';
 import { PreCreatedUserAttributes } from './pre-created-users.attributes';
 import { PreCreatedUser } from './pre-created-user.domain';
@@ -78,7 +74,12 @@ import { UpdateProfileDto } from './dto/update-profile.dto';
 import { isUUID } from 'class-validator';
 import { KeyServerUseCases } from '../keyserver/key-server.usecase';
 import { UserKeysEncryptVersions } from '../keyserver/key-server.domain';
-import { Kyber512 } from '../../lib/assymetric-encryption/kyber';
+import { AsymmetricEncryptionService } from '../../externals/asymmetric-encryption/asymmetric-encryption.service';
+import {
+  decryptMessageWithPrivateKey,
+  encryptMessageWithPublicKey,
+  generateNewKeys,
+} from '../../externals/asymmetric-encryption/openpgp';
 
 export class ReferralsNotAvailableError extends Error {
   constructor() {
@@ -151,6 +152,7 @@ export class UserUseCases {
     private readonly mailLimitRepository: SequelizeMailLimitRepository,
     private readonly featureLimitRepository: SequelizeFeatureLimitsRepository,
     private readonly keyServerUseCases: KeyServerUseCases,
+    private readonly asymmetricEncryptionService: AsymmetricEncryptionService,
   ) {}
 
   findByEmail(email: User['email']): Promise<User | null> {
@@ -453,6 +455,7 @@ export class UserUseCases {
     email: string,
     newUserUuid: string,
     newPublicKey: string,
+    newPublicKyberKey?: string,
   ) {
     const preCreatedUser =
       await this.preCreatedUserRepository.findByUsername(email);
@@ -462,30 +465,38 @@ export class UserUseCases {
     }
 
     const defaultPass = this.configService.get('users.preCreatedPassword');
-    const { privateKey: encPrivateKey } = preCreatedUser;
+    const { privateKey: encPrivateKey, privateKyberKey: encPrivateKyberKey } =
+      preCreatedUser;
+
     const privateKey = aes.decrypt(encPrivateKey, defaultPass);
+    const privateKyberKey = aes.decrypt(encPrivateKyberKey, defaultPass);
 
     const invites = await this.sharingRepository.getInvitesBySharedwith(
       preCreatedUser.uuid,
     );
 
     for (const invite of invites) {
-      const decryptedEncryptionKey = await decryptMessageWithPrivateKey({
-        encryptedMessage: Buffer.from(invite.encryptionKey, 'base64').toString(
-          'binary',
-        ),
-        privateKeyInBase64: privateKey,
-      });
+      const { encryptionKey } = invite;
 
-      const newEncryptedEncryptionKey = await encryptMessageWithPublicKey({
-        message: decryptedEncryptionKey.toString(),
-        publicKeyInBase64: newPublicKey,
-      });
+      const decryptedEncryptionKey =
+        await this.asymmetricEncryptionService.hybridDecryptMessageWithPrivateKey(
+          {
+            encryptedMessageInBase64: encryptionKey,
+            privateKeyInBase64: privateKey,
+            privateKyberKeyInBase64: privateKyberKey,
+          },
+        );
 
-      invite.encryptionKey = Buffer.from(
-        newEncryptedEncryptionKey.toString(),
-        'binary',
-      ).toString('base64');
+      const newEncryptedEncryptionKey =
+        await this.asymmetricEncryptionService.hybridEncryptMessageWithPublicKey(
+          {
+            message: decryptedEncryptionKey.toString(),
+            publicKeyInBase64: newPublicKey,
+            publicKyberKeyBase64: newPublicKyberKey,
+          },
+        );
+
+      invite.encryptionKey = newEncryptedEncryptionKey;
 
       invite.sharedWith = newUserUuid;
     }
@@ -586,8 +597,8 @@ export class UserUseCases {
       salt: this.configService.get('secrets.magicSalt'),
     });
 
-    const Kem = new Kyber512();
-    const kyberKeys = await Kem.generateKeysInBase64();
+    const kyberKeys =
+      await this.asymmetricEncryptionService.generateKyberKeys();
 
     const encPrivateKyberKey = aes.encrypt(kyberKeys.privateKey, defaultPass, {
       iv: this.configService.get('secrets.magicIv'),
